backend/requirements.txt
django>=5.2,<5.3.0
djangorestframework>=3.16,<3.17
requests>=2.32.2,<2.33
djangorestframework-simplejwt>=5.5,<5.6
dj-rest-auth>=7.0,<7.1
psycopg2-binary>=2.9,<2.10